Major changes in back-end.
Moved back-end of do-test.php into a class. And called when submitting form. However there is an issue since the post method is not working properly for the moment. Will be resolved in the following days.

For do-test.php: Made some permanent changes in the input names and some temporary changes in css.